# openapi.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"getData","parameters":[],"responses":{"200":{"description":""}}}},"/api/submissions/submit":{"post":{"operationId":"createSubmission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubmissionDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubmissionResponseDto"}}}}},"tags":["submissions"]}},"/api/submissions/run":{"post":{"operationId":"createRun","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubmissionDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubmissionResponseDto"}}}}},"tags":["submissions"]}},"/api/submissions/status/run":{"get":{"operationId":"getRunStatus","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RunStatusResponseDto"}}}}},"tags":["submissions"]}},"/api/submissions/status/submit":{"get":{"operationId":"getSubmitStatus","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitStatusResponseDto"}}}}},"tags":["submissions"]}},"/api/submissions/user/{problemId}":{"get":{"operationId":"getSubmissionsByUserAndProblem","parameters":[{"name":"problemId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Submission"}}}}}},"tags":["submissions"]}},"/api/submissions/{submissionId}":{"get":{"operationId":"getSubmissionById","parameters":[{"name":"submissionId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitStatusResponseDto"}}}}},"tags":["submissions"]}},"/api/problems":{"post":{"operationId":"createProblem","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProblemDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorProblemDto"}}}}},"tags":["problems"]},"get":{"operationId":"getProblems","parameters":[{"name":"pageIndex","in":"query","required":true,"schema":{"type":"number"}},{"name":"pageSize","in":"query","required":true,"schema":{"type":"number"}},{"example":"ASC or DESC","required":false,"name":"order","in":"query","schema":{"$ref":"#/components/schemas/SortOrder"}},{"required":false,"name":"title","in":"query","schema":{"type":"string"}},{"required":false,"name":"authorId","in":"query","schema":{"type":"number"}},{"required":false,"name":"difficulty","in":"query","schema":{"$ref":"#/components/schemas/ProblemDifficulty"}},{"required":false,"name":"status","in":"query","schema":{"$ref":"#/components/schemas/ProblemStatus"}},{"required":false,"name":"tagIds","in":"query","schema":{"type":"array","items":{"type":"number"}}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AllProblemsDto"}}}}},"tags":["problems"]}},"/api/problems/{problemId}":{"get":{"operationId":"getProblem","parameters":[{"name":"problemId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProblemDto"}}}}},"tags":["problems"]},"put":{"operationId":"updateProblem","parameters":[{"name":"problemId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProblemDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorProblemDto"}}}}},"tags":["problems"]}},"/api/problems/author/{problemId}":{"get":{"operationId":"getProblemForAuthor","parameters":[{"name":"problemId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorProblemDto"}}}}},"tags":["problems"]}},"/api/problems/admin":{"get":{"operationId":"getProblemsForAdmin","parameters":[{"name":"pageIndex","in":"query","required":true,"schema":{"type":"number"}},{"name":"pageSize","in":"query","required":true,"schema":{"type":"number"}},{"example":"ASC or DESC","required":false,"name":"order","in":"query","schema":{"$ref":"#/components/schemas/SortOrder"}},{"required":false,"name":"title","in":"query","schema":{"type":"string"}},{"required":false,"name":"authorId","in":"query","schema":{"type":"number"}},{"required":false,"name":"difficulty","in":"query","schema":{"$ref":"#/components/schemas/ProblemDifficulty"}},{"required":false,"name":"status","in":"query","schema":{"$ref":"#/components/schemas/ProblemStatus"}},{"required":false,"name":"tagIds","in":"query","schema":{"type":"array","items":{"type":"number"}}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AllProblemsDto"}}}}},"tags":["problems"]}},"/api/problems/admin/{problemId}":{"get":{"operationId":"getProblemForAdmin","parameters":[{"name":"problemId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminProblemDto"}}}}},"tags":["problems"]}},"/api/problems/admin/{problemId}/change-status":{"put":{"operationId":"changeProblemStatus","parameters":[{"name":"problemId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeProblemStatusDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorProblemDto"}}}}},"tags":["problems"]}},"/api/problems/admin/{problemId}/add-testcases":{"post":{"operationId":"addTestCasesToProblem","parameters":[{"name":"problemId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTestCasesDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AdminProblemDto"}}}}},"tags":["problems"]}},"/api/tags":{"post":{"operationId":"createTag","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}}},"tags":["tags"]},"get":{"operationId":"getAllTags","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/"}}}}},"tags":["tags"]}},"/api/tags/{tagId}":{"get":{"operationId":"getTag","parameters":[{"name":"tagId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}}},"tags":["tags"]},"put":{"operationId":"updateTag","parameters":[{"name":"tagId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTagDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tag"}}}}},"tags":["tags"]}},"/api/auth/google":{"get":{"operationId":"googleAuth","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/google/callback":{"get":{"operationId":"googleAuthRedirect","parameters":[],"responses":{"200":{"description":""}},"tags":["auth"]}},"/api/auth/sign-up":{"post":{"operationId":"signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerificationEmailSentDto"}}}}},"tags":["auth"]}},"/api/auth/resend-verification-email":{"post":{"operationId":"resendVerificationEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResendVerificationEmailDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerificationEmailSentDto"}}}}},"tags":["auth"]}},"/api/auth/verify-email":{"post":{"operationId":"verifyEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignedInUserResponseDto"}}}}},"tags":["auth"]}},"/api/auth/sign-in":{"post":{"operationId":"signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignedInUserResponseDto"}}}}},"tags":["auth"]}},"/api/auth/request-sign-in-otp":{"post":{"operationId":"requestSignInOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestSignInWithOtpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessMessageDto"}}}}},"tags":["auth"]}},"/api/auth/sign-in-with-otp":{"post":{"operationId":"signInWithOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInWithOtpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignedInUserResponseDto"}}}}},"tags":["auth"]}},"/api/auth/sign-in-with-token":{"post":{"operationId":"signInWithToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyTokenDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignedInUserResponseDto"}}}}},"tags":["auth"]}},"/api/auth/forgot-password":{"post":{"operationId":"forgotPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessMessageDto"}}}}},"tags":["auth"]}},"/api/auth/reset-password":{"post":{"operationId":"resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["auth"]}},"/api/auth/refresh":{"get":{"operationId":"refreshSession","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshedSessionDto"}}}}},"tags":["auth"]}},"/api/auth/logout":{"post":{"operationId":"logout","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuccessMessageDto"}}}}},"tags":["auth"]}},"/api/auth/check-username":{"post":{"operationId":"checkUsernameAvailability","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckUsernameDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"boolean"}}}}},"tags":["auth"]}},"/api/users/whoami":{"get":{"operationId":"whoAmI","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]}},"/api/users/onboard":{"put":{"operationId":"onboard","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OnboardUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]}},"/api/users/profile/edit":{"put":{"operationId":"editProfile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EditProfileDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]}},"/api/users/admin/users":{"get":{"operationId":"getAllUsers","parameters":[{"name":"pageIndex","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AllUsersDto"}}}}},"tags":["users"]}},"/api/users/admin/{userId}/change-role":{"put":{"operationId":"changeUserRole","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangeUserRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["users"]}},"/api/solutions":{"post":{"operationId":"createSolution","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSolutionDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolutionDto"}}}}},"tags":["solutions"]},"get":{"operationId":"getAllSolutions","parameters":[{"name":"pageIndex","in":"query","required":true,"schema":{"type":"number"}},{"name":"pageSize","in":"query","required":true,"schema":{"type":"number"}},{"example":"ASC or DESC","required":false,"name":"order","in":"query","schema":{"$ref":"#/components/schemas/SortOrder"}},{"name":"problemId","in":"query","required":true,"schema":{"type":"number"}},{"name":"language","in":"query","required":true,"schema":{"$ref":"#/components/schemas/SupportedLanguages"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AllSolutionsDto"}}}}},"tags":["solutions"]}},"/api/solutions/{solutionId}":{"get":{"operationId":"getSolutionById","parameters":[{"name":"solutionId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolutionDto"}}}}},"tags":["solutions"]}}},"info":{"title":"Code Judge","description":"An API to do awesome things","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Languages":{"type":"number","enum":[]},"CreateSubmissionDto":{"type":"object","properties":{"problemId":{"type":"number"},"sourceCode":{"type":"string"},"language":{"$ref":"#/components/schemas/Languages"}},"required":["problemId","sourceCode","language"]},"CreateSubmissionResponseDto":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]},"Language":{"type":"string","enum":["c","cpp","java","js","go"]},"SubmissionState":{"type":"string","enum":["Pending"," Started","Running","Success","Error"]},"ResultDto":{"type":"object","properties":{"input":{"type":"string"},"output":{"type":"string"},"expectedOutput":{"type":"string"}},"required":["input","output","expectedOutput"]},"RunStatusResponseDto":{"type":"object","properties":{"sourceCode":{"type":"string"},"language":{"$ref":"#/components/schemas/Language"},"state":{"$ref":"#/components/schemas/SubmissionState"},"error":{"type":"string"},"result":{"type":"array","items":{"$ref":"#/components/schemas/ResultDto"}},"passed":{"type":"number"},"total":{"type":"number"}},"required":["sourceCode","language","state","error","result","passed","total"]},"SubmissionStatus":{"type":"string","enum":["Accepted","Wrong Answer","Rejected","Compile error","Execution error","Runtime error","Time limit exceeded","Memory limit exceeded","Unexpected error"]},"SubmitStatusResponseDto":{"type":"object","properties":{"sourceCode":{"type":"string"},"language":{"$ref":"#/components/schemas/Language"},"state":{"$ref":"#/components/schemas/SubmissionState"},"error":{"type":"string"},"passed":{"type":"number"},"total":{"type":"number"},"status":{"$ref":"#/components/schemas/SubmissionStatus"},"memory":{"type":"number"},"time":{"type":"number"},"createdAt":{"format":"date-time","type":"string"}},"required":["sourceCode","language","state","error","passed","total","status","memory","time","createdAt"]},"Submission":{"type":"object","properties":{"id":{"type":"integer"},"language":{"type":"string","enum":["c","cpp","java","js","go"]},"status":{"$ref":"#/components/schemas/SubmissionStatus"},"totalTestCases":{"type":"number"},"testCasesPassed":{"type":"number"},"time":{"type":"number"},"memory":{"type":"number"},"createdAt":{"format":"date-time","type":"string"}},"required":["id","language","status","totalTestCases","testCasesPassed","time","memory","createdAt"]},"ProblemDifficulty":{"type":"string","enum":["Easy","Medium","Hard"]},"TestCaseDto":{"type":"object","properties":{"input":{"type":"string"},"output":{"type":"string"}},"required":["input","output"]},"CreateProblemDto":{"type":"object","properties":{"title":{"type":"string"},"difficulty":{"$ref":"#/components/schemas/ProblemDifficulty"},"description":{"type":"string"},"tagIds":{"type":"array","items":{"type":"number"}},"testCases":{"type":"array","items":{"$ref":"#/components/schemas/TestCaseDto"}}},"required":["title","difficulty","description","testCases"]},"ProblemStatus":{"type":"string","enum":["unpublished","approved","rejected"]},"Tag":{"type":"object","properties":{"id":{"type":"integer"},"name":{"type":"string"}},"required":["id","name"]},"UserRole":{"type":"string","enum":["user","problem_admin","super_admin"]},"User":{"type":"object","properties":{"id":{"type":"integer"},"username":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"isEmailVerified":{"type":"boolean"},"hasOnboarded":{"type":"boolean"},"role":{"$ref":"#/components/schemas/UserRole"}},"required":["id","username","firstName","lastName","email","isEmailVerified","hasOnboarded","role"]},"TestCaseType":{"type":"string","enum":["example","actual"]},"TestCase":{"type":"object","properties":{"id":{"type":"integer"},"type":{"$ref":"#/components/schemas/TestCaseType"},"input":{"type":"string"},"output":{"type":"string"}},"required":["id","type","input","output"]},"AuthorProblemDto":{"type":"object","properties":{"id":{"type":"integer"},"title":{"type":"string"},"difficulty":{"$ref":"#/components/schemas/ProblemDifficulty"},"status":{"$ref":"#/components/schemas/ProblemStatus"},"hasPlatformTestCases":{"type":"boolean"},"remark":{"type":"string"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}},"author":{"$ref":"#/components/schemas/User"},"description":{"type":"string"},"exampleTestCases":{"type":"array","items":{"$ref":"#/components/schemas/TestCase"}}},"required":["id","title","difficulty","status","hasPlatformTestCases","remark","tags","author","description","exampleTestCases"]},"SortOrder":{"type":"string","enum":["ASC","DESC"]},"Problem":{"type":"object","properties":{"id":{"type":"integer"},"title":{"type":"string"},"difficulty":{"$ref":"#/components/schemas/ProblemDifficulty"},"status":{"$ref":"#/components/schemas/ProblemStatus"},"hasPlatformTestCases":{"type":"boolean"},"remark":{"type":"string"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}},"author":{"$ref":"#/components/schemas/User"}},"required":["id","title","difficulty","status","hasPlatformTestCases","remark","tags","author"]},"PaginationResultDto":{"type":"object","properties":{"pageIndex":{"type":"number"},"pageSize":{"type":"number"},"totalPages":{"type":"number"},"totalItems":{"type":"number"},"itemsOnPage":{"type":"number"},"isFirstPage":{"type":"boolean"},"isLastPage":{"type":"boolean"}},"required":["pageIndex","pageSize","totalPages","totalItems","itemsOnPage","isFirstPage","isLastPage"]},"AllProblemsDto":{"type":"object","properties":{"problems":{"type":"array","items":{"$ref":"#/components/schemas/Problem"}},"paginationMeta":{"$ref":"#/components/schemas/PaginationResultDto"}},"required":["problems","paginationMeta"]},"ProblemDto":{"type":"object","properties":{"id":{"type":"integer"},"title":{"type":"string"},"difficulty":{"$ref":"#/components/schemas/ProblemDifficulty"},"status":{"$ref":"#/components/schemas/ProblemStatus"},"hasPlatformTestCases":{"type":"boolean"},"remark":{"type":"string"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}},"author":{"$ref":"#/components/schemas/User"},"description":{"type":"string"},"exampleTestCases":{"type":"array","items":{"$ref":"#/components/schemas/TestCase"}}},"required":["id","title","difficulty","status","hasPlatformTestCases","remark","tags","author","description","exampleTestCases"]},"UpdateProblemDto":{"type":"object","properties":{"title":{"type":"string"},"difficulty":{"$ref":"#/components/schemas/ProblemDifficulty"},"description":{"type":"string"},"tagIds":{"type":"array","items":{"type":"number"}},"testCases":{"type":"array","items":{"$ref":"#/components/schemas/TestCaseDto"}}},"required":["title","difficulty","description","testCases"]},"AdminProblemDto":{"type":"object","properties":{"id":{"type":"integer"},"title":{"type":"string"},"difficulty":{"$ref":"#/components/schemas/ProblemDifficulty"},"status":{"$ref":"#/components/schemas/ProblemStatus"},"hasPlatformTestCases":{"type":"boolean"},"remark":{"type":"string"},"tags":{"type":"array","items":{"$ref":"#/components/schemas/Tag"}},"author":{"$ref":"#/components/schemas/User"},"description":{"type":"string"},"exampleTestCases":{"type":"array","items":{"$ref":"#/components/schemas/TestCaseDto"}},"actualTestCases":{"type":"array","items":{"$ref":"#/components/schemas/TestCaseDto"}}},"required":["id","title","difficulty","status","hasPlatformTestCases","remark","tags","author","description","exampleTestCases","actualTestCases"]},"ChangeProblemStatusDto":{"type":"object","properties":{"status":{"$ref":"#/components/schemas/ProblemStatus"},"remark":{"type":"string"}},"required":["status","remark"]},"CreateTestCasesDto":{"type":"object","properties":{"testCases":{"type":"array","items":{"$ref":"#/components/schemas/TestCaseDto"}}},"required":["testCases"]},"CreateTagDto":{"type":"object","properties":{"tagName":{"type":"string"}},"required":["tagName"]},"SignUpUserDto":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["username","email","password"]},"VerificationEmailSentDto":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"ResendVerificationEmailDto":{"type":"object","properties":{"userId":{"type":"number"},"email":{"type":"string"}},"required":["userId","email"]},"VerifyEmailDto":{"type":"object","properties":{"userId":{"type":"number"},"email":{"type":"string"},"otp":{"type":"string"}},"required":["userId","email","otp"]},"SignedInUserResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["accessToken","user"]},"SignInUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"RequestSignInWithOtpDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"SuccessMessageDto":{"type":"object","properties":{"message":{"type":"string"}},"required":["message"]},"SignInWithOtpDto":{"type":"object","properties":{"email":{"type":"string"},"otp":{"type":"string"}},"required":["email","otp"]},"VerifyTokenDto":{"type":"object","properties":{"verificationToken":{"type":"string"}},"required":["verificationToken"]},"ForgotPasswordDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"password":{"type":"string"},"verificationToken":{"type":"string"}},"required":["password","verificationToken"]},"RefreshedSessionDto":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"CheckUsernameDto":{"type":"object","properties":{"username":{"type":"string"}},"required":["username"]},"OnboardUserDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["firstName","lastName"]},"EditProfileDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"}},"required":["firstName","lastName"]},"AllUsersDto":{"type":"object","properties":{"users":{"type":"array","items":{"$ref":"#/components/schemas/User"}},"paginationMeta":{"$ref":"#/components/schemas/PaginationResultDto"}},"required":["users","paginationMeta"]},"ChangeUserRoleDto":{"type":"object","properties":{"role":{"$ref":"#/components/schemas/UserRole"}},"required":["role"]},"CreateSolutionDto":{"type":"object","properties":{"problemId":{"type":"number"},"description":{"type":"string"},"language":{"$ref":"#/components/schemas/Languages"}},"required":["problemId","description","language"]},"SolutionDto":{"type":"object","properties":{"id":{"type":"integer"},"language":{"$ref":"#/components/schemas/Languages"},"createdAt":{"format":"date-time","type":"string"},"description":{"type":"string"}},"required":["id","language","createdAt","description"]},"SupportedLanguages":{"type":"number","enum":[]},"Solution":{"type":"object","properties":{"id":{"type":"integer"},"language":{"$ref":"#/components/schemas/Languages"},"createdAt":{"format":"date-time","type":"string"}},"required":["id","language","createdAt"]},"AllSolutionsDto":{"type":"object","properties":{"solutions":{"type":"array","items":{"$ref":"#/components/schemas/Solution"}},"paginationMeta":{"$ref":"#/components/schemas/PaginationResultDto"}},"required":["solutions","paginationMeta"]}}}}